refactor: add ComparisonCellRenderer for consistent comparison cell rendering
